Time: 3 ms (54.31%), Space: 9.2 MB (49.38%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.c
@@ -1,5 +1,5 @@
 int reverse(int x){
-    int rev = 0;
+    long long rev = 0;
 
     while (x != 0) {
         int digit = x%10;
